change integer to double

+++ src/Calculator.java
@@ -1,5 +1,5 @@
 public class Calculator {
-    public Integer calculate(Integer first, Integer second, String operation) {
+    public Double calculate(Double first, Integer second, String operation) {
 
         switch (operation) {
             case "+":
@@ -12,7 +12,7 @@ public class Calculator {
                 return first * second;
             default:
                 System.out.println("Такая операция не поддерживается.");
-                return -1;
+                return Double.valueOf(-1);
         }
     }
